refactor: MicState.tsx 파일명 수정
파스칼 케이스로 이름이 되어있어서 수정합니다.
