+++ frontend/src/components/Header/components/stateMicrophone.tsx
